zjmp corrected

+++ vm_dir/bytes.c
@@ -6,23 +6,23 @@
 /*   By: jleblond <jleblond@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/08 14:45:55 by jleblond          #+#    #+#             */
-/*   Updated: 2020/02/09 15:31:11 by jleblond         ###   ########.fr       */
+/*   Updated: 2020/02/17 20:48:23 by jleblond         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vm.h"
 
-uint32_t 	read_bytes(uint8_t *s, uint8_t nb_of_bytes)
+int32_t 	read_bytes(uint8_t *s, uint8_t nb_of_bytes)
 {
-	uint32_t 	ret;
+	int32_t 	ret;
 
 	ret = 0;
 	if (nb_of_bytes == 4)
 		ret = s[0] << 24 | s[1] << 16 | s[2] << 8 | s[3];
 	else if (nb_of_bytes == 2)
-		ret = s[0] << 8 | s[1];
+		ret = (int16_t)(s[0] << 8 | s[1]);
 	else if (nb_of_bytes == 1)
-		ret = s[0];
+		ret = (int8_t)(s[0]);
 	else
 	{
 		printf("nb_of_bytes [%d] wrong\n", nb_of_bytes);
